Move update ID bookkeeping to the Receiver
The counter for update ID is updated when a message is received, rather
than when during processing of the message. We ignore all malformed
messages.

Downside: After receiving a message, it is completely our responsibility
to ensure that it is stored properly so that any replies that are not
sent properly can be resent.